[Renderer] Warm up the renderer properly. (#50408)
